assignment1-dev-bhatia-1/webserver/src/requirements.txt
pyramid==1.10
mysql-connector-python
pyramid-jinja2==2.7
python-dotenv
